2022.05.09

+++ src/bronze/twoRule/Main.java
@@ -6,18 +6,16 @@ public class Main {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
-		int c = sc.nextInt();
+		
+		double a = sc.nextDouble();
+		double b = sc.nextDouble();
+		double c = sc.nextDouble();
 		
 		if (b > c) {
 			System.out.println((int)((a * b) / c));
 		} else {
 			System.out.println((int)((a * c) / b));
 		}
-		
-		
-		
 		sc.close();
 	}
